silence some sign-compare warnings

+++ src/filter/tutorial/tutorial.cpp
@@ -64,7 +64,7 @@ public:
         // We'll use a calculation that looks expensive here.
         float f, h;
         int tempVal;
-        for (int i = 0; i < lookupTable.size(); i++) {
+        for (size_t i = 0; i < lookupTable.size(); i++) {
             f = i / (float)std::numeric_limits<uint8_t>::max(); // Normalize to [0,1]
             h = f/5;
             f = 2*f - 1; // Stretch to [-1,1]
@@ -129,8 +129,8 @@ public:
         if (m_pointerMethod == 0) {
             uint8_t *in_pointer = (uint8_t *) in;
             uint8_t *out_pointer = (uint8_t *) out;
-            for (int x = 0; x < width; x++) {
-                for (int y = 0; y < height; y++) {
+            for (unsigned int x = 0; x < width; x++) {
+                for (unsigned int y = 0; y < height; y++) {
 
                     // Apply the parabola to the R channel with a single lookup
                     *out_pointer++ = lookupTable[*in_pointer++];
@@ -148,7 +148,7 @@ public:
             // This method takes only 80% of the time if only processing the R channel,
             // and 90% of the time with additionally the G channel,
             // compared to the above solution using uint8_t pointers.
-            for (int px = 0; px < width*height; px++) {
+            for (unsigned int px = 0; px < width*height; px++) {
                 out[px] =
                             // Parabola to Red channel
                             lookupTable[(in[px] & 0xFF)]
